pyproject.toml
[project]
name = "django-redis-postgres-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-redis>=6.0.0",
    "djangorestframework>=3.17.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
